add-persistent-storage - renaming.
